refactor(web): split oversized runtime modules
